Add extra blank newlines at the beginning of the PR template.
If we start the PR template description on the first line, then the
author has to manually add couple of newline. For instance, when Copilot
generate a summary, it does not add a newline and we end up with PR
starting at the end of the last line of the summary instead of a
new line. It's a bit messy.

Signed-off-by: Grégoire Payen de La Garanderie <gregoire.payen.de.la.garanderie@intel.com>