Use uv for dependency management (#1866)
This pull request modernizes the build and dependency management for the
project by migrating from `setup.py` and requirements files to a fully
declarative `pyproject.toml` configuration and switching the recommended
workflow to use the `uv` tool for dependency management and testing. It
also updates Python version support to require 3.10+ (3.9 EOLs in
October and was causing some problems). The C++ compiler is now also
optional, as a fallback was added to use python implementation in case
it is not installed.



<!--
Please add a summary of changes. You may use Copilot to auto-generate
the PR description but please consider including any other relevant
facts which Copilot may be unaware of (such as design choices and
testing procedure).

Add references to the relevant issues and pull requests if any like so:

Resolves #111 and #222.
Depends on #1000 (for series of dependent commits).
-->

### Checklist
<!-- Put an 'x' in all the boxes that apply -->
- [ ] I have added tests to cover my changes or documented any manual
tests.
- [ ] I have added the description of my changes into
[CHANGELOG](https://github.com/open-edge-platform/datumaro/blob/develop/CHANGELOG.md).
- [ ] I have updated the
[documentation](https://github.com/open-edge-platform/datumaro/tree/develop/docs)
accordingly

---------

Signed-off-by: Albert van Houten <albert.van.houten@intel.com>