Tiling support (#1905)
This PR adds a new tiling transform. It also introduces a unified
mechanism to represent transforms like tiling, converters, and
filtering. Transforms can be chained and evaluated lazy.

The tiling transform can be used like so:
```
    # Create tiling config
    config = TilingConfig(tile_width=100, tile_height=100, overlap_x=20, overlap_y=20)

    # Create and apply tiling transform
    tiling_transform = create_tiling_transform(config)
    tiled_dataset = sample_dataset.transform(tiling_transform)
```

<!-- Contributing guide:
https://github.com/open-edge-platform/datumaro/blob/develop/CONTRIBUTING.md
-->

<!--
Please add a summary of changes. You may use Copilot to auto-generate
the PR description but please consider including any other relevant
facts which Copilot may be unaware of (such as design choices and
testing procedure).

Add references to the relevant issues and pull requests if any like so:

Resolves #111 and #222.
Depends on #1000 (for series of dependent commits).
-->

### Checklist
<!-- Put an 'x' in all the boxes that apply -->
- [ ] I have added tests to cover my changes or documented any manual
tests.
- [ ] I have updated the
[documentation](https://github.com/open-edge-platform/datumaro/tree/develop/docs)
accordingly

---------

Signed-off-by: Grégoire Payen de La Garanderie <gregoire.payen.de.la.garanderie@intel.com>
Signed-off-by: Albert van Houten <albert.van.houten@intel.com>
Co-authored-by: Albert van Houten <albert.van.houten@intel.com>