Add support for Python 3.14 (#1962)
### Summary

This PR enables compatibility with the latest Python 3.14, while
dropping support for the old Python 3.10. Removing 3.10 was necessary
because there's no numpy version which supports 3.10 and 3.14 at the
same time. This PR also upgrades some dependency packages to a version
compatible with 3.14; for breaking changes (e.g. in `shapely`) I have
updated the source code.
CI was also updated to test both the minimum (3.11) and maximum (3.14)
supported versions.

Resolves #1960 

### Checklist
- [x] I have added tests to cover my changes or documented any manual
tests.
- [x] I have updated the
[documentation](https://github.com/open-edge-platform/datumaro/tree/develop/docs)
accordingly

---------

Signed-off-by: Leonardo Lai <leonardo.lai@intel.com>