Pin dependencies and restore py310 support (#1968)
In a previous PR #1962 I had removed support for Python 3.10 because
incompatible with the latest numpy. However, I realized that by
loosening the numpy version constraint a little, we can still keep
support for Python 3.10: it is restored by this PR. At the same time, I
reordered the dependencies alphabetically and pinned their version with
`~=` instead of `>=` to prevent breaking changes (only for dependencies
that follow semantic versioning).

Signed-off-by: Leonardo Lai <leonardo.lai@intel.com>