Coco Dataset support (#1963)
This pull request introduces comprehensive support for the COCO dataset
format in the `datumaro.experimental` module, including new data
structures, IO utilities, and field types. The changes add the ability
to load and save datasets in COCO format, define a schema for COCO
samples, and extend the field system to handle captions, numerics and
booleans. The numeric and boolean fields are generalized fields which
can be used for multiple purposes. For example, numeric can be used to
present a score for a prediction or as an identifier for images.
Additionally, the enum for semantics has been replaced with a string to
allow more flexibility. ScoreField has been removed, as NumericField can
now be used to represent these values.

**COCO Datasets** 
Logic to load and save coco datasets was added in a new folder:
"experimental/data_formats/coco". There are 2 main files:

**io.py** Contains logic to both load and save COCO datasets, including
those from different version years.
**sample.py** Defines a COCO sample that is able to hold all annotation
data, including polygons, bboxes, keypoints and captions.
 

**Performance**
Tiny coco dataset with 100 train and 100 val images loads in ~0.17
seconds. (https://github.com/lizhogn/tiny_coco_dataset)

<!-- Contributing guide:
https://github.com/open-edge-platform/datumaro/blob/develop/contributing.md
-->

<!--
Please add a summary of changes. You may use Copilot to auto-generate
the PR description but please consider including any other relevant
facts which Copilot may be unaware of (such as design choices and
testing procedure).

Add references to the relevant issues and pull requests if any like so:

Resolves #111 and #222.
Depends on #1000 (for series of dependent commits).
-->

### Checklist
<!-- Put an 'x' in all the boxes that apply -->
- [ ] I have added tests to cover my changes or documented any manual
tests.
- [ ] I have updated the
[documentation](https://github.com/open-edge-platform/datumaro/tree/develop/docs)
accordingly

---------

Signed-off-by: Albert van Houten <albert.van.houten@intel.com>